Crate axerrno

Error code definition used by ArceOS.

It provides two error types and the corresponding result types:

AxError and AxResult: A generic error type similar to std::io::ErrorKind.

LinuxError and LinuxResult: Linux specific error codes defined in errno.h. It can be converted from AxError.

Macros

ax_err

Convenience method to construct an Err(AxError) type while printing a warning message.

ax_err_type

Convenience method to construct an AxError type while printing a warning message.

ensure

Ensure a condition is true. If it is not, return from the function with an error.

Enums

AxError

The error type used by ArceOS.

LinuxError

Linux specific error codes defined in errno.h.

Type Aliases

AxResult

A specialized Result type with AxError as the error type.

LinuxResult

A specialized Result type with LinuxError as the error type.